Update boilerplate skin
These updates should bring the theme closer to the current HAUS identity and brand. Also added some comments to the `application.scss` partial explaining why that partial exists and that it should probably be deleted when starting a project. Updated the logo. Cleaned up styles – @extends at the top, @includes at the bottom of scss definitions.